Add check for second monitor & prioritize it.
Someone with a second monitor thought it was a good idea to let the game open on it.

And now I'm stuck with the consequences. My second monitor died a few days ago. And this piece of shit of an app doesn't work if there was no second monitor.

Added a check for a second monitor